requirements.txt
streamlit>=1.30.0
requests>=2.31.0
pandas>=2.1.0
numpy>=1.25.0
scipy>=1.11.0
python-dotenv>=1.0.0
plotly>=5.18.0
